Split out HullType and auxilliary into ShipHull.{h,cpp}
