+++ Empire/ProductionQueue.cpp
@@ -3,8 +3,8 @@
 #include "Empire.h"
 #include "../universe/BuildingType.h"
 #include "../universe/Condition.h"
+#include "../universe/ShipHull.h"
 #include "../universe/ShipPart.h"
-#include "../universe/ShipPartHull.h"
 #include "../universe/ShipDesign.h"
 #include "../universe/ValueRef.h"
 #include "../util/AppInterface.h"
