+++ UI/BuildDesignatorWnd.cpp
@@ -18,8 +18,8 @@
 #include "../Empire/EmpireManager.h"
 #include "../universe/BuildingType.h"
 #include "../universe/ShipDesign.h"
+#include "../universe/ShipHull.h"
 #include "../universe/ShipPart.h"
-#include "../universe/ShipPartHull.h"
 #include "../universe/Conditions.h"
 #include "../universe/Enums.h"
 #include "../universe/ValueRef.h"
