+++ msvc2017/Common/Common.vcxproj
@@ -224,6 +224,7 @@
     <ClInclude Include="..\..\universe\ResourceCenter.h" />
     <ClInclude Include="..\..\universe\Ship.h" />
     <ClInclude Include="..\..\universe\ShipDesign.h" />
+    <ClInclude Include="..\..\universe\ShipHull.h" />
     <ClInclude Include="..\..\universe\ShipPart.h" />
     <ClInclude Include="..\..\universe\Special.h" />
     <ClInclude Include="..\..\universe\Species.h" />
@@ -301,6 +302,7 @@
     <ClCompile Include="..\..\universe\ResourceCenter.cpp" />
     <ClCompile Include="..\..\universe\Ship.cpp" />
     <ClCompile Include="..\..\universe\ShipDesign.cpp" />
+    <ClCompile Include="..\..\universe\ShipHull.cpp" />
     <ClCompile Include="..\..\universe\ShipPart.cpp" />
     <ClCompile Include="..\..\universe\Special.cpp" />
     <ClCompile Include="..\..\universe\Species.cpp" />
