+++ parse/ShipHullsParser.cpp
@@ -10,7 +10,7 @@
 #include "CommonParamsParser.h"
 
 #include "../universe/Condition.h"
-#include "../universe/ShipPartHull.h"
+#include "../universe/ShipHull.h"
 #include "../universe/ValueRef.h"
 
 #include <boost/spirit/include/phoenix.hpp>
