+++ python/UniverseWrapper.cpp
@@ -5,7 +5,7 @@
 #include "../universe/Ship.h"
 #include "../universe/ShipDesign.h"
 #include "../universe/ShipPart.h"
-#include "../universe/ShipPartHull.h"
+#include "../universe/ShipHull.h"
 #include "../universe/Building.h"
 #include "../universe/BuildingType.h"
 #include "../universe/ResourceCenter.h"
